# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Usul API",
    "description": "API for accessing Islamic texts, books, authors, and AI-powered chat functionality",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.usul.ai",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/search/": {
      "get": {
        "tags": ["Search"],
        "summary": "Global search",
        "description": "Search across books, authors, and genres with content search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query"
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "ar"]
            },
            "description": "Locale for localized content"
          }
        ],
        "responses": {
          "200": {
            "description": "Global search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/search/empires": {
      "get": {
        "tags": ["Search"],
        "summary": "Search empires",
        "description": "Search for empires with optional sorting",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Search query"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of results to return"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "default": 1 },
            "description": "Page number for pagination"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "relevance",
                "texts-asc",
                "texts-desc",
                "authors-asc",
                "authors-desc",
                "alphabetical-asc",
                "alphabetical-desc"
              ]
            },
            "description": "Sort order for results"
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["en", "ar"] },
            "description": "Locale for localized content"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpireSearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/genre": {
      "get": {
        "tags": ["Genres"],
        "summary": "Get all genres",
        "description": "Get all genres with optional filters",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "ar"]
            },
            "description": "Locale for localized content"
          },
          {
            "name": "bookIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of book IDs to filter by"
          },
          {
            "name": "yearRange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated year range (e.g., '1000,1500')"
          },
          {
            "name": "authorId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Author ID to filter by"
          },
          {
            "name": "regionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Region ID to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Genres retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Genre"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/region": {
      "get": {
        "tags": ["Regions"],
        "summary": "Get all regions",
        "description": "Get all geographical regions",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "ar"]
            },
            "description": "Locale for localized content"
          }
        ],
        "responses": {
          "200": {
            "description": "Regions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Region"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/empire": {
      "get": {
        "tags": ["Empires"],
        "summary": "Get all empires",
        "description": "Get all empires with optional filters",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "ar"]
            },
            "description": "Locale for localized content"
          },
          {
            "name": "yearRange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated year range (e.g., '1000,1500')"
          },
          {
            "name": "genreId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Genre ID to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Empires retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Empire"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/empire/{slug}": {
      "get": {
        "tags": ["Empires"],
        "summary": "Get empire by slug",
        "description": "Retrieve empire information using its slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The slug of the empire"
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "ar"]
            },
            "description": "Locale for localized content"
          }
        ],
        "responses": {
          "200": {
            "description": "Empire retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Empire"
                }
              }
            }
          },
          "404": {
            "description": "Empire not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/collections/": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get all collections",
        "description": "Get all collections for the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Collections retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Collection"
                      }
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Collections"],
        "summary": "Create a collection",
        "description": "Create a new book collection",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Name of the collection"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the collection"
                  },
                  "slug": {
                    "type": "string",
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "minLength": 1,
                    "description": "URL-friendly slug (lowercase alphanumeric with hyphens)"
                  },
                  "visibility": {
                    "type": "string",
                    "enum": ["PUBLIC", "UNLISTED"],
                    "description": "Visibility of the collection"
                  }
                },
                "required": ["name", "description", "slug", "visibility"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Collection"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - collection with this slug already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/advancedGenre": {
      "get": {
        "tags": ["Advanced Genres"],
        "summary": "Get all advanced genres",
        "description": "Get all advanced genres with optional filters",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["en", "ar"] },
            "description": "Locale for localized content"
          },
          {
            "name": "bookIds",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Comma-separated list of book IDs to filter by"
          },
          {
            "name": "yearRange",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Comma-separated year range (e.g., '1000,1500')"
          },
          {
            "name": "authorId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Author ID to filter by"
          },
          {
            "name": "regionId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Region ID to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Advanced genres retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AdvancedGenre" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvancedGenre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the advanced genre"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "name": { "type": "string", "description": "Primary display name" },
          "secondaryName": {
            "type": "string",
            "description": "Secondary display name"
          },
          "numberOfBooks": {
            "type": "integer",
            "description": "Associated books count (if available)"
          }
        },
        "required": ["id", "slug", "name"]
      },
...
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the book"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "primaryName": {
            "type": "string",
            "description": "Primary name of the book"
          },
          "secondaryName": {
            "type": "string",
            "description": "Secondary name of the book"
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "year": {
            "type": "integer",
            "description": "Year of publication"
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookVersion"
            }
          }
        },
        "required": ["id", "slug", "primaryName", "author"]
      },
      "BookVersion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the version"
          },
          "value": {
            "type": "string",
            "description": "Version identifier"
          },
          "source": {
            "type": "string",
            "description": "Source of the version"
          },
          "aiSupported": {
            "type": "boolean",
            "description": "Whether AI is supported for this version"
          },
          "keywordSupported": {
            "type": "boolean",
            "description": "Whether keyword search is supported"
          }
        },
        "required": ["id", "value", "source"]
      },
      "Author": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the author"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "primaryName": {
            "type": "string",
            "description": "Primary name of the author"
          },
          "secondaryName": {
            "type": "string",
            "description": "Secondary name of the author"
          },
          "year": {
            "type": "integer",
            "description": "Year of birth/death"
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Location"
            }
          }
        },
        "required": ["id", "slug", "primaryName"]
      },
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the location"
          },
          "name": {
            "type": "string",
            "description": "Name of the location"
          },
          "region": {
            "$ref": "#/components/schemas/Region"
          }
        },
        "required": ["id", "name"]
      },
      "Region": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the region"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "name": {
            "type": "string",
            "description": "Name of the region"
          }
        },
        "required": ["id", "slug", "name"]
      },
      "Empire": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the empire"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "name": {
            "type": "string",
            "description": "Primary name of the empire"
          },
          "secondaryName": {
            "type": "string",
            "description": "Secondary name of the empire"
          },
          "overview": {
            "type": "string",
            "description": "Overview description of the empire"
          },
          "numberOfAuthors": {
            "type": "integer",
            "description": "Number of authors associated with this empire"
          },
          "numberOfBooks": {
            "type": "integer",
            "description": "Number of books associated with this empire"
          }
        },
        "required": ["id", "slug", "numberOfAuthors", "numberOfBooks"]
      },
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the genre"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "name": {
            "type": "string",
            "description": "Name of the genre"
          },
          "color": {
            "type": "string",
            "description": "Color for UI display"
          },
          "pattern": {
            "type": "integer",
            "description": "Pattern number for UI display"
          }
        },
        "required": ["id", "slug", "name"]
      },
      "Collection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the collection"
          },
          "name": {
            "type": "string",
            "description": "Name of the collection"
          },
          "description": {
            "type": "string",
            "description": "Description of the collection"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly slug identifier"
          },
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "UNLISTED"],
            "description": "Visibility of the collection"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who owns the collection"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            },
            "description": "Books in the collection"
          }
        },
        "required": ["id", "name", "slug", "visibility", "userId"]
      },
...
      "GlobalSearchResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "results": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "versionId": {
                      "type": "string"
                    },
                    "book": {
                      "$ref": "#/components/schemas/Book"
                    },
                    "node": {
                      "type": "object",
                      "description": "Search result node"
                    }
                  }
                }
              }
            }
          },
          "books": {
            "type": "object",
            "properties": {
              "hits": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Book"
                }
              },
              "found": {
                "type": "integer"
              }
            }
          },
          "authors": {
            "type": "object",
            "properties": {
              "hits": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Author"
                }
              },
              "found": {
                "type": "integer"
              }
            }
          },
          "genres": {
            "type": "object",
            "properties": {
              "hits": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Genre"
                }
              },
              "found": {
                "type": "integer"
              }
            }
          }
        },
        "required": ["content", "books", "authors", "genres"]
      },
...
      "EmpireSearchResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "object",
            "properties": {
              "hits": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Empire" }
              },
              "found": { "type": "integer" }
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": { "type": "integer" },
              "perPage": { "type": "integer" },
              "totalPages": { "type": "integer" }
            }
          }
        },
        "required": ["results", "pagination"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["status", "message"]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      }
    }
  }
}