# src/dto/author.dto.ts
import { PathLocale } from '@/lib/locale';
import { getPrimaryLocalizedText, getSecondaryLocalizedText } from '@/lib/localization';
import { Author, AuthorBio, AuthorOtherNames, AuthorPrimaryName } from '@prisma/client';
import { getLocationById } from '@/services/location';

export type AuthorDto = ReturnType<typeof makeAuthorDto>;

export const makeAuthorDto = (
  author: Author & {
    primaryNameTranslations: AuthorPrimaryName[];
    otherNameTranslations: AuthorOtherNames[];
    bioTranslations: AuthorBio[];
    locations: { id: string }[];
    empires: { id: string }[];
  },
  locale: PathLocale,
  {
    includeLocations = false,
  }: {
    includeLocations?: boolean;
  } = {},
) => {
  const primaryName = getPrimaryLocalizedText(author.primaryNameTranslations, locale);
  const otherNames = getPrimaryLocalizedText(author.otherNameTranslations, locale) ?? [];

  return {
    id: author.id,
    slug: author.slug,
    year: author.year,
    numberOfBooks: author.numberOfBooks,
    primaryName:
      locale === 'en' && author.transliteration ? author.transliteration : primaryName,
    otherNames:
      locale === 'en' && author.otherNameTransliterations?.length > 0
        ? author.otherNameTransliterations
        : otherNames,

    secondaryName: getSecondaryLocalizedText(author.primaryNameTranslations, locale),
    secondaryOtherNames:
      getSecondaryLocalizedText(author.otherNameTranslations, locale) ?? [],

    bio: getPrimaryLocalizedText(author.bioTranslations, locale),

    empires: author.empires.map(empire => ({ id: empire.id })),

    ...(includeLocations && {
      locations: author.locations.map(location =>
        getLocationById(location.id, locale, { includeRegion: true }),
      ),
    }),
  };
};